+++ pilad/conn.go
@@ -263,6 +263,12 @@ func (c *Conn) stackHandler(params *map[string]string) http.Handler {
 			return
 
 		case r.Method == "POST":
+			_ = r.ParseForm()
+			if _, ok := r.Form["rotate"]; ok {
+				c.rotateStackHandler(w, r, stack)
+				return
+			}
+
 			c.checkMaxStackSize(c.addElementStackHandler)(w, r, stack)
 			return
 
@@ -347,16 +353,32 @@ func (c *Conn) fullStackHandler(w http.ResponseWriter, r *http.Request, stack *p
 	w.Write(fullStackHandlerResponse)
 }
 
+// rotateStackHandler rotates the bottommost element of the Stack
+// to the top.
+func (c *Conn) rotateStackHandler(w http.ResponseWriter, r *http.Request, stack *pila.Stack) {
+	ok := stack.Rotate()
+	if !ok {
+		log.Println(r.Method, r.URL, http.StatusNoContent)
+		w.WriteHeader(http.StatusNoContent)
+		return
+	}
+	stack.Update(c.opDate)
+
+	var element pila.Element
+	element.Value = stack.Peek()
+
+	log.Println(r.Method, r.URL, http.StatusOK, element.Value)
+	w.Header().Set("Content-Type", "application/json")
+
+	// Do not check error as we consider our element
+	// suitable for a JSON encoding.
+	b, _ := element.ToJSON()
+	w.Write(b)
+}
+
 // addElementStackHandler adds an element into a Stack and returns 200 and the element.
 // It can be as a PUSH or a BASE operation.
 func (c *Conn) addElementStackHandler(w http.ResponseWriter, r *http.Request, stack *pila.Stack) {
-	if r.Body == nil {
-		log.Println(r.Method, r.URL, http.StatusBadRequest,
-			"no element provided")
-		w.WriteHeader(http.StatusBadRequest)
-		return
-	}
-
 	var element pila.Element
 	err := element.Decode(r.Body)
 	if err != nil {